farmarket_ai/requirements.txt
fastapi
uvicorn
pandas
httpx
python-dotenv
